no revert fullscreen on retina displays
A "fix" was put in to place for Chrome version 16 that is very old
(Chrome is at version 49 now) and is causing issues on retina
screens (see https://github.com/johndyer/mediaelement/issues/1708)

This patchset removes the Chrome 16 "fix" in order to fix the new
issues it causes on these displays.

Also detects fullscreen on 'canplay' event instead of 'play' since
it's possible to click the 'fullscreen' icon without having first
played the video (in the event that an iframe-embedded video player
does not first detect the fullscreen mode, it will try to go full-
screen inside the iframe, even on browsers that support native-
native mode).