allow descendent nodes to be insulated from keyboard theft
mejs adds global keydown and click handlers that mark the player as
"hasFocus" any time any descendent node of the player gets clicked or
keydown-ed. and when it hasFocus, the player responds to keydown events.
this means if we mount an interactive node, e.g. an `<input>`, to a
player's DOM tree, e.g. its `.mejs-layers`, then trying to e.g. type a
space in that input will toggle play/pause on the player rather than
putting a space character in the input.

this commit enables the ability to thwart that keydown event stealing.
if you assign the class ".mejs-keyboard-insulator" to a DOM node, it and
all of its descendents will be insulated from that theft.