Hide overlay for all browsers (refs ARC-4600)
Chrome 72 changed the default policy for autoplay by disallowing playing
video without direct interaction with the video's domain happening
first. That is, if video controls exist in domain A and the video plays
in an iframe with domain B, then unless the user interacts with domain B
first, the controls in domain A cannot be used to play the video.

In most cases, mediaelement is rendering an overlay that captures clicks
on every part of the iframe. This is okay for videos that share a domain
but in the case of Youtube embeds, the youtube domain will never
receive a direct click. This change hides that overlay in the same way
that was done in a previous mitigation for iOS and Android, but now for
all browsers.